simplify deployment on remote servers
Set parameter that defines the host of the backend API only
during development mode. For deployment on remote servers or localhost,
we assume that frontend and backend are deployed on the same host, at
the same base URL and port. We then derive hostname and port
based on the URL that the frontend is access at.